+++ sarathi/model_executor/attention/flashinfer_attention_wrapper.py
@@ -8,8 +8,6 @@ from sarathi.core.datatypes.sequence import SequenceMetadata
 from sarathi.metrics.constants import OperationMetrics
 from sarathi.model_executor.attention.base_attention_wrapper import BaseAttentionWrapper
 
-KVCache = Union[Tuple[torch.Tensor, torch.Tensor], torch.Tensor]
-
 class FlashinferAttentionWrapper(BaseAttentionWrapper):
 
     def __init__(
@@ -51,28 +49,6 @@ class FlashinferAttentionWrapper(BaseAttentionWrapper):
 
     def to_int_tensor(self, data: List[int]) -> torch.Tensor:
         return torch.tensor(data, dtype=torch.int32, device="cuda")
-    
-    def init_gpu_cache(self, num_gpu_blocks: int) -> None:
-        gpu_cache: List[torch.Tensor] = []
-        self.num_gpu_blocks = num_gpu_blocks
-
-        for _ in range(self.num_layers):
-            gpu_blocks = self.get_cache_block(
-                self.num_gpu_blocks, dtype=self.dtype, device="cuda"
-            )
-            gpu_cache.append(gpu_blocks)
-        
-        self.gpu_cache = gpu_cache
-
-    def get_cache_block(self, num_blocks: int, **kwargs) -> torch.Tensor:
-        return torch.randn(
-            num_blocks,
-            2,
-            self.block_size,
-            self.num_kv_heads,
-            self.head_dim,
-            **kwargs,
-        )
 
     def get_cache_block_size(self) -> int:
         key_cache_block = self.block_size * self.num_kv_heads * self.head_dim
