+++ sarathi/model_executor/attention/no_op_attention_wrapper.py
@@ -18,15 +18,7 @@ class NoOpAttentionWrapper(BaseAttentionWrapper):
     ):
         self.device = device
 
-    def init_gpu_cache(self, num_gpu_blocks: int) -> None:
-        pass
-
     def get_cache_block_size(self) -> int:
-        pass
-    
-    def get_cache_block(
-        self, num_blocks: int, **kwargs
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
         pass
 
     def begin_forward(
